requirements.txt
# Libraires that are required for the project

# Web framework
fastapi
uvicorn

# Object storage
minio
boto3

# Database
sqlalchemy
python-dotenv
pandas
psycopg2-binary

# Config / env
python-dotenv
pydantic
pydantic-settings

# Data processing
pandas
scikit-learn

# ML
catboost
joblib

# MLflow experiment tracking (Sprint 2)
mlflow
pika==1.3.2

# FastAPI file upload support
python-multipart

celery>=5.3.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-celery
